Move button in info screen
