+++ components/info/impl/src/main/res/values/colors.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <color name="update_green">#2ED832</color>
     <color name="battery_backgroud">#8D8E92</color>
     <color name="device_info_release">#2ED832</color>
 </resources>
