+++ components/updater/fonts/build.gradle.kts
@@ -1,0 +1,6 @@
+plugins {
+    androidLibrary
+}
+
+dependencies {
+}
