+++ components/updater/fonts/src/main/res/font/flipper.ttf
