+++ components/updater/fonts/src/main/res/font/flipper_bold.ttf
